Web Dev/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
Jinja2==3.1.2
stripe==7.0.0
reportlab==4.0.4
gunicorn==21.2.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
mypy==1.5.1
python-jose==3.3.0
bcrypt==4.0.1
requests==2.31.0
aiohttp==3.8.5
asyncio==3.4.3
aiofiles==23.1.0
python-multipart==0.0.6
email-validator==2.0.0
Pillow==10.0.0
qrcode==7.4.2
weasyprint==60.1
markdown==3.4.4
bleach==6.0.0
python-slugify==8.0.1
humanize==4.7.0
pytz==2023.3 